pyproject.toml
[tool.poetry]
name = "msgapp"
version = "0.1.0"
description = "Declarative message processing applications"
authors = ["Adrian Garcia Badaracco <dev@adriangb.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
anyio = "^3.6.1"
pydantic = {version = "^1.9.1", optional = true }
google-cloud-pubsub = { version = "^2.13.4", optional = true }
aiobotocore = { version = "^2.3.4", optional = true }

[tool.poetry.extras]
sqs = ["aiobotocore"]
pubsub = ["google-cloud-pubsub"]
json = ["pydantic"]
dev = [
    "aiobotocore",
    "google-cloud-pubsub",
    "pydantic"
]

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
pytest = "*"
pre-commit = "*"
pytest-cov = "*"
trio = "*"
isort = "*"
flake8 = "*"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.coverage.run]
branch = true
source = ["msgapp", "tests"]
dynamic_context = "test_function"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    '\.\.\.',
    "raise NotImplementedError",
    "TYPE_CHECKING",
    "__repr__",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
files = "msgapp/**/*.py,tests/**/*.py"
show_error_codes = true
pretty = true
warn_unreachable = true
strict = true
